--- api/client/config.go
@@ -1,83 +1,0 @@
-/*
-Copyright 2021 Gravitational, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package client
-
-import (
-	"time"
-
-	"github.com/gravitational/teleport/api/defaults"
-	"github.com/gravitational/trace"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/keepalive"
-)
-
-// Config contains configuration of the client
-type Config struct {
-	// Addrs is a list of teleport auth/proxy server addresses to dial.
-	Addrs []string
-	// Credentials are a list of credentials to use when attempting
-	// to form a connection from client to server.
-	Credentials []Credentials
-	// Dialer is a custom dialer used to dial a server. If set, Dialer
-	// takes precedence over all other connection options.
-	Dialer ContextDialer
-	// DialOpts define options for dialing the client connection.
-	DialOpts []grpc.DialOption
-	// DialInBackground specifies to dial the connection in the background
-	// rather than blocking until the connection is up. A predefined Dialer
-	// or an auth server address must be provided.
-	DialInBackground bool
-	// DialTimeout defines how long to attempt dialing before timing out.
-	DialTimeout time.Duration
-	// KeepAlivePeriod defines period between keep alives.
-	KeepAlivePeriod time.Duration
-	// KeepAliveCount specifies the amount of missed keep alives
-	// to wait for before declaring the connection as broken.
-	KeepAliveCount int
-	// The web proxy uses a self-signed TLS certificate by default, which
-	// requires this field to be set. If the web proxy was provided with
-	// signed TLS certificates, this field should not be set.
-	InsecureAddressDiscovery bool
-}
-
-// CheckAndSetDefaults checks and sets default config values.
-func (c *Config) CheckAndSetDefaults() error {
-	if len(c.Credentials) == 0 {
-		return trace.BadParameter("missing connection credentials")
-	}
-
-	if c.KeepAlivePeriod == 0 {
-		c.KeepAlivePeriod = defaults.ServerKeepAliveTTL
-	}
-	if c.KeepAliveCount == 0 {
-		c.KeepAliveCount = defaults.KeepAliveCountMax
-	}
-	if c.DialTimeout == 0 {
-		c.DialTimeout = defaults.DefaultDialTimeout
-	}
-
-	c.DialOpts = append(c.DialOpts, grpc.WithKeepaliveParams(keepalive.ClientParameters{
-		Time:                c.KeepAlivePeriod,
-		Timeout:             c.KeepAlivePeriod * time.Duration(c.KeepAliveCount),
-		PermitWithoutStream: true,
-	}))
-	if !c.DialInBackground {
-		c.DialOpts = append(c.DialOpts, grpc.WithBlock())
-	}
-
-	return nil
-}
